Add tests to run as Docker run command
Previously, tests were run as part of the Docker build process. This is valuable since it can allow the container to fail building if tests fail BUT does not allow for a lot of flexibility.

This commit chanes the Dockerfile so that the tests are copied to a final container so they can be run. This required the creation of an entrypoint to catch the `test` command (during `docker run`).

The tests can no be run with:

```
make tests
```

Additionally, running tests during build time is still available by setting `--build-arg TEST_ON_BUILD=true`